requirements.txt
pymysql
flask
cryptography
pycrypto
mysqlclient
flask_migrate
flask_script
xmltodict
pyjwt
flask_session